Administrasjon: Ikke mulig å legge til underområder i program. Ved refresh forsvinner de man la til.
