+++ SharePointFramework/ProgramWebParts/src/components/ProgramAdministration/AddProjectDialog/AddProjectDialog.tsx
@@ -13,7 +13,7 @@ import * as strings from 'ProgramWebPartsStrings'
 import React, { FC, useContext } from 'react'
 import { columns } from '../columns'
 import { ProgramAdministrationContext } from '../context'
-import { ListHeaderSearch } from '../ListHeaderSearch'
+import { ListHeaderSearch } from '../ListHeaderSearch/ListHeaderSearch'
 import { ADD_CHILD_PROJECTS, TOGGLE_ADD_PROJECT_DIALOG } from '../reducer'
 import styles from './AddProjectDialog.module.scss'
 import { useAddProjectDialog } from './useAddProjectDialog'
