+++ SharePointFramework/ProgramWebParts/src/components/ProgramAdministration/AddProjectDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './AddProjectDialog'
