+++ SharePointFramework/ProgramWebParts/src/components/ProgramAdministration/ProgramAdministration.tsx
@@ -3,9 +3,9 @@ import { isEmpty } from '@microsoft/sp-lodash-subset'
 import * as strings from 'ProgramWebPartsStrings'
 import { WebPartTitle } from 'pp365-shared-library'
 import React, { FC } from 'react'
-import { AddProjectDialog } from './AddProjectDialog'
-import { Commands } from './Commands'
-import { ListHeaderSearch } from './ListHeaderSearch'
+import { AddProjectDialog } from './AddProjectDialog/AddProjectDialog'
+import { Commands } from './Commands/Commands'
+import { ListHeaderSearch } from './ListHeaderSearch/ListHeaderSearch'
 import styles from './ProgramAdministration.module.scss'
 import { columns } from './columns'
 import { ProgramAdministrationContext } from './context'
