+++ SharePointFramework/ProgramWebParts/src/components/ProgramAdministration/index.ts
@@ -1,0 +1,1 @@
+export * from './ProgramAdministration'
