+++ SharePointFramework/ProgramWebParts/src/components/ProgramAdministration/useProgramAdministration.ts
@@ -16,7 +16,7 @@ export const useProgramAdministration = (props: IProgramAdministrationProps) => 
   useEffect(() => {
     props.dataAdapter.project.getProjectInformationData().then((properties) => {
       Promise.all([
-        props.dataAdapter.fetchChildProjects(),
+        props.dataAdapter.fetchChildProjects(properties.fieldValues.id),
         props.dataAdapter.checkProjectAdminPermissions(
           ProjectAdminPermission.ChildProjectsAdmin,
           properties.fieldValues
