+++ SharePointFramework/ProgramWebParts/src/webparts/baseProgramWebPart/index.ts
@@ -6,7 +6,7 @@ import { createSpfiInstance } from 'pp365-shared-library'
 import { IHubSite } from 'pp365-shared-library/lib/interfaces'
 import { ComponentClass, FC, ReactElement, createElement } from 'react'
 import { render } from 'react-dom'
-import { SPDataAdapter } from '../../data'
+import { SPDataAdapter } from 'data/SPDataAdapter'
 import { IBaseProgramWebPartProps } from './types'
 
 export abstract class BaseProgramWebPart<
